requirements.txt
asttokens==2.0.5
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
docutils==0.17.1
executing==0.6.0
idna==2.10
importlib-metadata==4.0.1
keyring==23.0.1
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
pexpect==4.8.0
Pillow==8.2.0
pkginfo==1.7.0
ptyprocess==0.7.0
pure-eval==0.2.1
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.60.0
twine==3.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
webencodings==0.5.1
zipp==3.4.1
